backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
pydantic==2.10.4
pydantic-settings==2.7.1
httpx==0.28.1
apscheduler==3.10.4
python-dotenv==1.0.1
psycopg2-binary==2.9.9
scikit-learn>=1.4.0
